Replace remaining references to `tasks` with `steps`.
