+++ lib/plugins/Step/AbstractStep.js
@@ -162,7 +162,7 @@ class AbstractStep extends events.EventEmitter {
    * Returns a JSON representation of the plugin, which is used automatically with JSON.stringify.
    * List of JSONified properties can be overwridden with an array in 'this._jsonAttributes'.
    *
-   * @return {object} - A JSON representation of this task.
+   * @return {object} - A JSON representation of this step.
    */
   toJSON() {
     var output = {};
@@ -182,7 +182,7 @@ class AbstractStep extends events.EventEmitter {
   }
 
   /**
-   * Return a description string for the task. By default returns an empty string
+   * Return a description string for the step. By default returns an empty string
    * @return {String} - The description of the plugin.
    */
   description() {
@@ -280,14 +280,14 @@ class AbstractStep extends events.EventEmitter {
   }
 
   /**
-   * Call to emit an event that's a status update for the task
+   * Call to emit an event that's a status update for the step
    * @param {Object} status - The status object (state, action, description)
    * @param {string} [context] - Optional context value to use. If not set, defaults to a sensible value.
    */
   updateStatus(status, context) {
     context = context || `${this.plugin}/${this.name}`;
     status.description = status.description || this.description();
-    status.task = this;
+    status.step = this;
     this.emit('update', context, status, this);
   }
 }
